requirements.txt
dvc>=3.42.0
optuna>=4.0.0
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
joblib>=1.0.0
pyyaml>=5.4.0
